Added DutyPlan
DutyPlan now works:
DutyPlans are created automatically, if necessary, when a user logs in. Users are added depending on their score and the number of services as well as the Selected Available Days. If several users can be added, a random decision is made. Given FavPartners are also taken into account.